Add `snforge new` command
