+++ crates/forge/tests/e2e/running.rs
@@ -1,7 +1,7 @@
 use super::common::runner::{
     get_current_branch, get_remote_url, runner, setup_package, snforge_test_bin_path, test_runner,
 };
-use assert_fs::fixture::{FileWriteStr, PathChild, PathCopy};
+use assert_fs::fixture::{FileTouch, FileWriteStr, PathChild, PathCopy};
 use assert_fs::TempDir;
 use camino::Utf8PathBuf;
 use forge::scarb::config::SCARB_MANIFEST_TEMPLATE_CONTENT;
@@ -11,6 +11,7 @@ use shared::test_utils::output_assert::assert_stdout_contains;
 use snapbox::assert_matches;
 use snapbox::cmd::Command as SnapboxCommand;
 use std::ffi::OsString;
+use std::path::PathBuf;
 use std::{env, fs, iter, path::Path, str::FromStr};
 use test_utils::{get_local_snforge_std_absolute_path, tempdir_with_tool_versions};
 use toml_edit::{value, DocumentMut, Formatted, InlineTable, Item, Value};
@@ -696,7 +697,61 @@ fn init_new_project() {
         .assert()
         .success();
 
-    validate_init(&temp, false);
+    validate_init(&temp.join("test_name"), false);
+}
+
+#[test]
+fn create_new_project_dir_not_exist() {
+    let temp = tempdir_with_tool_versions().unwrap();
+    let project_path = temp.join("new").join("project");
+
+    runner(&temp)
+        .args(["new", "--name", "test_name"])
+        .arg(&project_path)
+        .env("DEV_DISABLE_SNFORGE_STD_DEPENDENCY", "true")
+        .assert()
+        .success();
+
+    validate_init(&project_path, false);
+}
+
+#[test]
+fn create_new_project_dir_not_empty() {
+    let temp = tempdir_with_tool_versions().unwrap();
+    temp.child("empty.txt").touch().unwrap();
+
+    let output = runner(&temp)
+        .args(["new", "--name", "test_name"])
+        .arg(temp.path())
+        .assert()
+        .code(2);
+
+    assert_stdout_contains(
+        output,
+        indoc!(
+            r"
+                [ERROR] The provided path [..] points to a non-empty directory. If you wish to create a project in this directory, please use the `--overwrite` flag to proceed.
+            "
+        ),
+    );
+}
+
+#[test]
+fn create_new_project_dir_exists_and_empty() {
+    let temp = tempdir_with_tool_versions().unwrap();
+    let project_path = temp.join("new").join("project");
+
+    fs::create_dir_all(&project_path).unwrap();
+    assert!(project_path.exists());
+
+    runner(&temp)
+        .args(["new", "--name", "test_name"])
+        .arg(&project_path)
+        .env("DEV_DISABLE_SNFORGE_STD_DEPENDENCY", "true")
+        .assert()
+        .success();
+
+    validate_init(&project_path, false);
 }
 
 #[test]
@@ -714,7 +769,7 @@ fn init_new_project_from_scarb() {
         .assert()
         .success();
 
-    validate_init(&temp, true);
+    validate_init(&temp.join("test_name"), true);
 }
 
 pub fn append_to_path_var(path: &Path) -> OsString {
@@ -724,8 +779,8 @@ pub fn append_to_path_var(path: &Path) -> OsString {
     env::join_paths(script_path.chain(other_paths)).unwrap()
 }
 
-fn validate_init(temp: &TempDir, validate_snforge_std: bool) {
-    let manifest_path = temp.join("test_name/Scarb.toml");
+fn validate_init(project_path: &PathBuf, validate_snforge_std: bool) {
+    let manifest_path = project_path.join("Scarb.toml");
     let scarb_toml = fs::read_to_string(manifest_path.clone()).unwrap();
 
     let snforge_std_assert = if validate_snforge_std {
@@ -784,8 +839,8 @@ fn validate_init(temp: &TempDir, validate_snforge_std: bool) {
 
     std::fs::write(manifest_path, scarb_toml.to_string()).unwrap();
 
-    let output = test_runner(temp)
-        .current_dir(temp.child(Path::new("test_name")))
+    let output = test_runner(&TempDir::new().unwrap())
+        .current_dir(project_path)
         .assert()
         .success();
 
@@ -812,14 +867,15 @@ fn test_init_project_with_custom_snforge_dependency_git() {
     let temp = tempdir_with_tool_versions().unwrap();
 
     runner(&temp)
-        .args(["init", "test_name"])
+        .args(["new", "test_name"])
         .env("DEV_DISABLE_SNFORGE_STD_DEPENDENCY", "true")
         .assert()
         .success();
 
-    let manifest_path = temp.child("test_name/Scarb.toml");
+    let project_path = temp.join("test_name");
+    let manifest_path = project_path.join("Scarb.toml");
 
-    let scarb_toml = std::fs::read_to_string(manifest_path.path()).unwrap();
+    let scarb_toml = std::fs::read_to_string(&manifest_path).unwrap();
     let mut scarb_toml = DocumentMut::from_str(&scarb_toml).unwrap();
 
     let dependencies = scarb_toml
@@ -838,10 +894,10 @@ fn test_init_project_with_custom_snforge_dependency_git() {
     dependencies.remove("snforge_std");
     dependencies.insert("snforge_std", Item::Value(Value::InlineTable(snforge_std)));
 
-    std::fs::write(manifest_path.path(), scarb_toml.to_string()).unwrap();
+    std::fs::write(&manifest_path, scarb_toml.to_string()).unwrap();
 
     let output = test_runner(&temp)
-        .current_dir(temp.child(Path::new("test_name")))
+        .current_dir(&project_path)
         .assert()
         .success();
